pyproject.toml
[project]
name = "project-api-data-orchestrator"
version = "0.1.0"
description = ""
authors = [
    {name = "trighele",email = "trighele@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.119.1,<0.120.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]

[tool.poetry]
packages = [{include = "project_api_data_orchestrator", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
